:truck: move certif course test to certif eval test context
