+++ api/tests/unit/domain/models/CertificationCenter_test.js
@@ -1,4 +1,4 @@
-import { domainBuilder, expect } from '../../../test-helper.js';
+import { domainBuilder, expect } from '../../../../test-helper.js';
 
 describe('Unit | Domain | Models | CertificationCenter', function () {
   describe('#isSco', function () {
